fix(tests): clear clippy -D warnings reds in feature-gated test suites
Two pre-existing lints, both in test files, both outside anything the
default CI clippy matrix compiled until now:

- resource_estimate.rs: top-level `ColorPrimaries` import was only used
  inside the hdr-experimental-gated test, so the default-features
  `cargo clippy --workspace --all-targets -- -D warnings` CI job failed
  with unused_imports — the sole red on the last 5 main CI runs. Scope
  the import into the gated test.
- cll_measure.rs: unnecessary `as usize` cast (bytes_per_pixel already
  returns usize), caught now that the hdr-experimental suite is being
  linted/run.

Verified: cargo clippy --workspace --all-targets -- -D warnings PASS;
cargo clippy -p zenpixels-convert --all-targets --features
hdr-experimental -- -D warnings PASS.

+++ zenpixels-convert/tests/cll_measure.rs
@@ -245,7 +245,7 @@ fn measure_max_rejects_integer_pixel_formats() {
     let desc = PixelDescriptor::RGB16
         .with_transfer(TransferFunction::Pq)
         .with_primaries(zenpixels::ColorPrimaries::Bt2020);
-    let data = alloc::vec![0u8; (desc.bytes_per_pixel() * 4) as usize];
+    let data = alloc::vec![0u8; desc.bytes_per_pixel() * 4];
     let buf = PixelBuffer::from_vec(data, 2, 2, desc).expect("u16 pq buf");
     let cll = ContentLightLevel::measure_max(
         buf.as_slice(),
